Moving InMemory Stream and System Descriptors to samza-core
