+++ samza-test/src/main/java/org/apache/samza/test/framework/TestRunner.java
@@ -60,9 +60,9 @@ import org.apache.samza.system.descriptors.StreamDescriptor;
 import org.apache.samza.system.inmemory.InMemorySystemFactory;
 import org.apache.samza.task.AsyncStreamTask;
 import org.apache.samza.task.StreamTask;
-import org.apache.samza.test.framework.system.descriptors.InMemoryInputDescriptor;
-import org.apache.samza.test.framework.system.descriptors.InMemoryOutputDescriptor;
-import org.apache.samza.test.framework.system.descriptors.InMemorySystemDescriptor;
+import org.apache.samza.system.inmemory.descriptors.InMemoryInputDescriptor;
+import org.apache.samza.system.inmemory.descriptors.InMemoryOutputDescriptor;
+import org.apache.samza.system.inmemory.descriptors.InMemorySystemDescriptor;
 import org.apache.samza.util.FileUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
