+++ samza-test/src/test/java/org/apache/samza/test/framework/AsyncStreamTaskIntegrationTest.java
@@ -29,9 +29,9 @@ import java.util.stream.Collectors;
 import org.apache.samza.SamzaException;
 import org.apache.samza.operators.KV;
 import org.apache.samza.serializers.NoOpSerde;
-import org.apache.samza.test.framework.system.descriptors.InMemoryInputDescriptor;
-import org.apache.samza.test.framework.system.descriptors.InMemoryOutputDescriptor;
-import org.apache.samza.test.framework.system.descriptors.InMemorySystemDescriptor;
+import org.apache.samza.system.inmemory.descriptors.InMemoryInputDescriptor;
+import org.apache.samza.system.inmemory.descriptors.InMemoryOutputDescriptor;
+import org.apache.samza.system.inmemory.descriptors.InMemorySystemDescriptor;
 import org.hamcrest.collection.IsIterableContainingInOrder;
 import org.junit.Assert;
 import org.junit.Test;
